requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
huggingface-hub==0.27.0
idna==3.10
jinja2==3.1.5
joblib==1.4.2
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
packaging==24.2
pydantic==2.10.4
pydantic-core==2.27.2
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
